1 species and 1 microhabitat now work properly
This should now be able to simulate a single species on an environment with a width of 1.
The disperse() function is changed so that the entire list of K matrices doesn't need to be inputted every time. Instead, it just uses the relevant K matrices (which should, theoretically, speed up computation.
The compete() function now allows for different types of competition (both "lottery" and "temp" based, while also using the interaction matrix A).
Other types of diversity (like gini-Simpsons) can be calculated with the diversityIndex() function (replacing the previous invSimp() function).